REFACTOR: Move most of main.go into its own library
Currently our main() is difficult to test: we can't use our usual
tooling from package main since they're marked as internal. We also
can't use things package main in end_to_end_tests since we cannot
import executable package as if they are a library. To cut through this
issue, this commit moves most of main() into a library package. A later
commit will add a new facade main() that calls this library.
